+++ dotnet/test/Snowflake44.Tests/Snowflake44GeneratorTests.vb
@@ -1,9 +1,9 @@
 Imports Microsoft.VisualStudio.TestTools.UnitTesting
 
-Namespace Uid64Tests
+Namespace Snowflake44Tests
 
   <TestClass>
-  Public Class Uid64GeneratorTests
+  Public Class Snowflake44GeneratorTests
 
     <TestMethod()>
     Public Sub GenerateUid_CalledQuicklyAndOften_DoesNotCreateDuplicateIds()
@@ -14,7 +14,7 @@ Namespace Uid64Tests
 
         previousUid = uid
 
-        uid = Uid64.Generate()
+        uid = Snowflake44.Generate()
         Assert.AreNotEqual(previousUid, uid)
 
       Next
@@ -25,13 +25,13 @@ Namespace Uid64Tests
     Public Sub DecodeDateTimeTest()
 
       Dim expectedDate1 As New Date(1973, 12, 9, 15, 33, 22)
-      Dim saltedTimestamp1 As Long = Uid64.EncodeDateTime(expectedDate1)
+      Dim saltedTimestamp1 As Long = Snowflake44.EncodeDateTime(expectedDate1)
 
       Dim expectedDate2 As Date = expectedDate1.AddMilliseconds(1)
-      Dim saltedTimestamp2 As Long = Uid64.EncodeDateTime(expectedDate2)
+      Dim saltedTimestamp2 As Long = Snowflake44.EncodeDateTime(expectedDate2)
 
-      Dim actualDate1 As Date = Uid64.DecodeDateTime(saltedTimestamp1)
-      Dim actualDate2 As Date = Uid64.DecodeDateTime(saltedTimestamp2)
+      Dim actualDate1 As Date = Snowflake44.DecodeDateTime(saltedTimestamp1)
+      Dim actualDate2 As Date = Snowflake44.DecodeDateTime(saltedTimestamp2)
 
       Assert.AreEqual(expectedDate1, actualDate1)
       Assert.AreEqual(expectedDate2, actualDate2)