+++ js/src/index.ts
@@ -1,1 +1,1 @@
-export * from "./uid64"
+export * from "./snowflake44"